Revise sound sample selection
From a discussion on Discord, I learned that variant selection
is also done with the n function, and so using the synth:index
notation in the wrong place can make the index be treated as
if one were calling n(index). By ensuring the sound selection
comes after the voicing() call, this problem is avoided.